requirements.txt
Flask==0.12.2
sysfs-gpio==0.2.2
Twisted==17.9.0
zope.interface==4.4.3
